requirements.txt
requests
PyYAML
fastapi
uvicorn
pytest
pytest-mock